added e2e tests
