+++ src/app/e2e/admin/resources.json
@@ -1,0 +1,78 @@
+{
+    "field": [{
+        "name" : "uri",
+        "transformers" : [{ "operation" : "COLUMN", "args" : [ { "name" : "column", "type" : "column", "value" : "id" } ] } ],
+        "label" : "URI",
+        "cover" : "collection",
+        "scheme" : "http://uri4uri.net/vocab"
+    }, {
+        "name" : "fullname",
+        "label" : "Full name",
+        "transformers" : [
+            { "operation" : "COLUMN", "args" : [ { "name" : "column", "type" : "column", "value" : "uid" } ] },
+            { "operation" : "UPPERCASE", "args" : [ ] }
+        ],
+        "cover" : "collection",
+        "scheme" : "http://www.w3.org/ns/person"
+    }, {
+        "name" : "email",
+        "label" : "Email",
+        "transformers" : [ { "operation" : "COLUMN", "args" : [ { "name" : "column", "type" : "column", "value" : "mail" } ] } ],
+        "cover" : "collection",
+        "scheme" : "http://uri4uri.net/vocab"
+    }, {
+        "name" : "movie",
+        "label" : "Movie",
+        "transformers" : [ { "operation" : "COLUMN", "args" : [ { "name" : "column", "type" : "column", "value" : "movie" } ] } ],
+        "cover" : "dataset",
+        "scheme" : "http://www.w3.org/ns/person"
+    }, {
+        "name" : "author",
+        "label" : "Author",
+        "transformers" : [ { "operation" : "COLUMN", "args" : [ { "name" : "column", "type" : "column", "value" : "author" } ] } ],
+        "cover" : "dataset",
+        "scheme" : "http://www.w3.org/ns/person"
+    }],
+    "publishedCharacteristic": [
+        { "movie": "LOTR" , "author": "Peter Jackson" }
+    ],
+    "publishedDataset": [
+        {
+            "uri" : "1",
+            "versions": [{
+                "fullname" : "PEREGRIN.TOOK",
+                "email" : "peregrin.took@shire.net"
+            }],
+            "removedAt": "2017-02-09T16:32:53.005Z"
+        },
+        {
+            "uri" : "2",
+            "versions": [{
+                "fullname" : "SAMSAGET.GAMGIE",
+                "email" : "samsaget.gamgie@shire.net"
+            }],
+            "removedAt": "2017-02-09T16:32:53.005Z"
+        },
+        {
+            "uri" : "3",
+            "versions": [{
+                "fullname" : "BILBON.BAGGINS",
+                "email" : "bilbon.saquet@shire.net"
+            }]
+        },
+        {
+            "uri" : "4",
+            "versions": [{
+                "fullname" : "FRODO.BAGGINS",
+                "email" : "frodo.saquet@shire.net"
+            }]
+        },
+        {
+            "uri" : "5",
+            "versions": [{
+                "fullname" : "MERIADOC.BRANDYBUCK",
+                "email" : "meriadoc.brandybuck@shire.net"
+            }]
+        }
+    ]
+}
