+++ src/app/e2e/home_published_julia.js
@@ -6,7 +6,7 @@ import { clear, loadFixtures } from '../../common/tests/fixtures';
 import fixtures from './home_published.json';
 import { elementIsClickable, elementValueIs } from '../../common/tests/conditions';
 
-describe.only('Home page with published data when logged as Julia', function homePublishedDataTests() {
+describe('Home page with published data when logged as Julia', function homePublishedDataTests() {
     this.timeout(10000);
     const DEFAULT_WAIT_TIMEOUT = 9000; // A bit less than mocha's timeout to get explicit errors from selenium
